refactor: move ResolvedMaterial to blobdb app (#9762)
* refactor: move ResolvedMaterial to blobdb app

* fix: undo accidental removal
